+++ src/components/SearchModal/SearchModal.js
@@ -165,15 +165,17 @@ const SearchModal = ({
     ? resources.accessTypes?.records[0]?.data.map(({ attributes }) => ({ value: attributes.name, label: attributes.name }))
     : [];
 
+  const getMappingData = ({ attributes, id, type }) => ({
+    ...attributes,
+    id,
+    type,
+    checked: !!currentSearchConfig.selectedItems.find(item => item.id === id),
+  });
+
   const getFormattedResourcesData = titleRecords => {
     return titleRecords.reduce((allResources, currentTitle) => {
       const titleResources = currentTitle.included;
-      const formattedTitleResources = titleResources.map(({ attributes, id, type }) => ({
-        ...attributes,
-        id,
-        type,
-        checked: !!currentSearchConfig.selectedItems.find(item => item.id === id),
-      }));
+      const formattedTitleResources = titleResources.map(getMappingData);
 
       return [
         ...allResources,
@@ -183,12 +185,7 @@ const SearchModal = ({
   };
 
   const getFormattedPackagesData = records => {
-    return records.map(({ attributes, id, type }) => ({
-      ...attributes,
-      id,
-      type,
-      checked: !!currentSearchConfig.selectedItems.find(item => item.id === id),
-    }));
+    return records.map(getMappingData);
   };
 
   const getFormattedListItems = () => {
@@ -492,15 +489,17 @@ const SearchModal = ({
     }
 
     return (
-      <div className={css.modalFooter}>
+      <div
+        className={css.modalFooter}
+        data-testid="modal-footer"
+      >
         <Button
-          data-test-find-package-title-cancel
           onClick={closeModal}
           marginBottom0
         >
           <FormattedMessage id="stripes-components.cancel" />
         </Button>
-        <div data-test-titles-selected-count>
+        <div>
           <FormattedMessage
             id="ui-plugin-find-package-title.resultsPane.totalSelected.count"
             values={{
@@ -509,7 +508,6 @@ const SearchModal = ({
           />
         </div>
         <Button
-          data-test-find-package-title-save
           marginBottom0
           buttonStyle="primary"
           onClick={handleSave}
@@ -531,6 +529,7 @@ const SearchModal = ({
       onClose={closeModal}
       size="large"
       id="find-package-title-modal"
+      data-testid="find-package-title-modal"
       footer={renderModalFooter()}
     >
       <Paneset static isRoot>