+++ test/bigtest/interactors/application.js
@@ -1,8 +1,0 @@
-import { interactor, isPresent } from '@bigtest/interactor';
-
-@interactor class ApplicationInteractor {
-  static defaultScope = '#ModuleContainer';
-  pluginNotFound = isPresent('[data-test-no-plugin-available]');
-}
-
-export default ApplicationInteractor;